+++ docs/source/CHANGELOG.md
@@ -48,5 +48,6 @@
 
 ### 1.4.0 (in progress)
 
+- Generalize custom configuration parameters
 - Improve generality of the app test manager configurations and fixup docstrings
 - Set `TESTING` attribute to `False` as default for configurations